server/requirements.txt
fastapi
uvicorn
pydantic
SQLAlchemy>=2.0
psycopg[binary]>=3.2
requests>=2.32
# Frontend 
 npm run dev
#  backend
 node index.js
#  AI reccomendation
 /Users/rocky/Documents/pace-web
 source server/.venv/bin/activate
 uvicorn server.main:app --reload --port 8000
#  Local Ai Engine(Ollamma)
brew install ollama
ollama serve
# in a new terminal:
ollama pull llama3.1:8b
# openai
python-dotenv
python-docx
requests